Rename details package to showdetails
