+++ app/src/main/java/me/banes/chris/tivi/inject/AppComponent.kt
@@ -21,9 +21,9 @@ import dagger.android.AndroidInjector
 import dagger.android.support.AndroidSupportInjectionModule
 import me.banes.chris.tivi.TiviApplication
 import me.banes.chris.tivi.data.DatabaseModule
-import me.banes.chris.tivi.details.ShowDetailsBuilder
 import me.banes.chris.tivi.home.HomeBuilder
 import me.banes.chris.tivi.jobs.JobsCreator
+import me.banes.chris.tivi.showdetails.ShowDetailsBuilder
 import me.banes.chris.tivi.tmdb.TmdbModule
 import me.banes.chris.tivi.trakt.TraktModule
 import javax.inject.Singleton
