+++ app/src/main/res/layout/activity_show_details.xml
@@ -20,7 +20,7 @@
     android:id="@+id/details_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.banes.chris.tivi.details.ShowDetailsActivity">
+    tools:context="me.banes.chris.tivi.showdetails.ShowDetailsActivity">
 
     <FrameLayout
         android:id="@+id/details_content"
