+++ app/src/main/res/layout/fragment_show_details.xml
@@ -76,6 +76,6 @@
         android:layout_marginStart="16dp"
         android:elevation="3dp"
         android:transitionName="poster"
-        app:layout_behavior="me.banes.chris.tivi.details.DetailPosterBehavior" />
+        app:layout_behavior="me.banes.chris.tivi.showdetails.details.DetailPosterBehavior" />
 
 </android.support.design.widget.CoordinatorLayout>
